server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "ocrmac>=0.1.12",
    "uvicorn>=0.42.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "isort>=8.0.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
